refactor(ocr): inline OCR template models/serializers/views into core
Move OcrTemplate/OcrTemplateZone into documents/models.py, their serializers
into documents/serialisers.py, and OcrTemplateViewSet into documents/views.py;
delete the standalone *_ocr_templates.py modules and update all importers.
Keeps the feature in paperless' conventional module layout for upstreaming.
